+++ eCommerce-backend/src/test/java/com/app/domain/review/controllers/members/MemberReviewControllerTest.java
@@ -86,7 +86,7 @@ public class MemberReviewControllerTest {
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andExpect(content().contentType(MediaType.APPLICATION_JSON))
-                .andExpect(jsonPath("$.id", is(reviewDTO.id())))
+                .andExpect(jsonPath("$.id", is(reviewDTO.id().intValue())))
                 .andExpect(jsonPath("$.author", notNullValue()))
                 .andExpect(jsonPath("$.rating", is(reviewDTO.rating())))
                 .andExpect(jsonPath("$.comment", notNullValue()));
@@ -167,7 +167,7 @@ public class MemberReviewControllerTest {
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andExpect(content().contentType(MediaType.APPLICATION_JSON))
-                .andExpect(jsonPath("$.id", is(reviewDTO.id())))
+                .andExpect(jsonPath("$.id", is(reviewDTO.id().intValue())))
                 .andExpect(jsonPath("$.author", notNullValue()))
                 .andExpect(jsonPath("$.rating", is(reviewDTO.rating())))
                 .andExpect(jsonPath("$.comment", notNullValue()));
